Split unknown_066AA0.c into unknown_06B2B0.c and moved the rest of the rodata
