+++ asm/non_matchings/unknown_066AA0/func_8006DCF8.s
@@ -1,3 +1,8 @@
+.late_rodata
+glabel jpt_800E71D8
+.word L8006DF70, L8006E00C, L8006E054, L8006E0E0, L8006DEE8, L8006E0E0, L8006E0E0, L8006E0E0, L8006E0E0, L8006E0E0, L8006E0E0, L8006E0E0, L8006E0E0, L8006DF00
+
+.text
 glabel func_8006DCF8
 /* 06E8F8 8006DCF8 3C018012 */  lui   $at, %hi(gIsPaused) # $at, 0x8012
 /* 06E8FC 8006DCFC 3C0E8012 */  lui   $t6, %hi(D_80123514) # $t6, 0x8012
@@ -121,18 +126,20 @@ glabel func_8006DCF8
 /* 06EAC4 8006DEC4 2DC1000E */  sltiu $at, $t6, 0xe
 /* 06EAC8 8006DEC8 10200085 */  beqz  $at, .L8006E0E0
 /* 06EACC 8006DECC 000E7080 */   sll   $t6, $t6, 2
-/* 06EAD0 8006DED0 3C01800E */  lui   $at, %hi(D_800E71D8) # $at, 0x800e
+/* 06EAD0 8006DED0 3C01800E */  lui   $at, %hi(jpt_800E71D8) # $at, 0x800e
 /* 06EAD4 8006DED4 002E0821 */  addu  $at, $at, $t6
-/* 06EAD8 8006DED8 8C2E71D8 */  lw    $t6, %lo(D_800E71D8)($at)
+/* 06EAD8 8006DED8 8C2E71D8 */  lw    $t6, %lo(jpt_800E71D8)($at)
 /* 06EADC 8006DEDC 00000000 */  nop   
 /* 06EAE0 8006DEE0 01C00008 */  jr    $t6
 /* 06EAE4 8006DEE4 00000000 */   nop   
+glabel L8006DEE8
 /* 06EAE8 8006DEE8 2404000F */  li    $a0, 15
 /* 06EAEC 8006DEEC 2405FFFF */  li    $a1, -1
 /* 06EAF0 8006DEF0 0C01B68A */  jal   func_8006DA28
 /* 06EAF4 8006DEF4 24060001 */   li    $a2, 1
 /* 06EAF8 8006DEF8 100000F8 */  b     .L8006E2DC
 /* 06EAFC 8006DEFC 8FBF0024 */   lw    $ra, 0x24($sp)
+glabel L8006DF00
 /* 06EB00 8006DF00 3C088012 */  lui   $t0, %hi(D_801234F4) # $t0, 0x8012
 /* 06EB04 8006DF04 3C098012 */  lui   $t1, %hi(D_80123504) # $t1, 0x8012
 /* 06EB08 8006DF08 3C0B8012 */  lui   $t3, %hi(D_80123508) # $t3, 0x8012
@@ -161,6 +168,7 @@ glabel func_8006DCF8
 /* 06EB64 8006DF64 00402025 */   move  $a0, $v0
 /* 06EB68 8006DF68 100000DC */  b     .L8006E2DC
 /* 06EB6C 8006DF6C 8FBF0024 */   lw    $ra, 0x24($sp)
+glabel L8006DF70
 /* 06EB70 8006DF70 3C058012 */  lui   $a1, %hi(D_80121250) # $a1, 0x8012
 /* 06EB74 8006DF74 24A51250 */  addiu $a1, %lo(D_80121250) # addiu $a1, $a1, 0x1250
 /* 06EB78 8006DF78 80B80000 */  lb    $t8, ($a1)
@@ -202,6 +210,7 @@ glabel func_8006DCF8
 /* 06EC00 8006E000 00402025 */   move  $a0, $v0
 /* 06EC04 8006E004 100000B5 */  b     .L8006E2DC
 /* 06EC08 8006E008 8FBF0024 */   lw    $ra, 0x24($sp)
+glabel L8006E00C
 /* 06EC0C 8006E00C 3C018012 */  lui   $at, %hi(D_801234EC) # $at, 0x8012
 /* 06EC10 8006E010 AC2034EC */  sw    $zero, %lo(D_801234EC)($at)
 /* 06EC14 8006E014 3C038012 */  lui   $v1, %hi(D_80123500) # $v1, 0x8012
@@ -220,6 +229,7 @@ glabel func_8006DCF8
 /* 06EC48 8006E048 00000000 */   nop   
 /* 06EC4C 8006E04C 100000A3 */  b     .L8006E2DC
 /* 06EC50 8006E050 8FBF0024 */   lw    $ra, 0x24($sp)
+glabel L8006E054
 /* 06EC54 8006E054 3C058012 */  lui   $a1, %hi(D_80121250) # $a1, 0x8012
 /* 06EC58 8006E058 3C018012 */  lui   $at, %hi(D_801234EC) # $at, 0x8012
 /* 06EC5C 8006E05C 24A51250 */  addiu $a1, %lo(D_80121250) # addiu $a1, $a1, 0x1250
@@ -256,6 +266,7 @@ glabel func_8006DCF8
 /* 06ECD8 8006E0D8 10000080 */  b     .L8006E2DC
 /* 06ECDC 8006E0DC 8FBF0024 */   lw    $ra, 0x24($sp)
 .L8006E0E0:
+glabel L8006E0E0
 /* 06ECE0 8006E0E0 00002025 */  move  $a0, $zero
 /* 06ECE4 8006E0E4 2405FFFF */  li    $a1, -1
 /* 06ECE8 8006E0E8 0C01B68A */  jal   func_8006DA28
