+++ asm/unknown_070110.s
@@ -1,5 +1,5 @@
 /* The comment below is needed for this file to be picked up by generate_ld */
-/* RAM_POS: 0x80065F10 */
+/* RAM_POS: 0x8006F510 */
 
 .section .text
